Getting question refactor to work
